Slider Library V:0.1
First installment, Allows for a slider to be moved to a specific value, given that you provide the slider as a WebElement and some characteristics about the slider such as Max, min, incriment and a webDriver. Then once you give it the amount you want it to be moved to it will move it there. hopefully speeding up testing through automation.